Actualización: Clientes y Mascotas (Corrección errores)

+++ PetFry/Articulo.cs
@@ -9,21 +9,21 @@ namespace PetFry
     public abstract class Articulo
     {
         private string _nombre;
-        private int _valor;
+        private decimal _valor;
         private string _notas;
 
         public string Nombre { get => _nombre; set => _nombre = value; }
-        public int Valor { get => _valor; set => _valor = value; }
+        public decimal Valor { get => _valor; set => _valor = value; }
         public string Notas { get => _notas; set => _notas = value; }
 
-        public Articulo(string nombre, int valor, string notas)
+        public Articulo(string nombre, decimal valor, string notas)
         {
             Nombre = nombre;
             Valor = valor;
             Notas = notas;
         }
 
-        public Articulo(string nombre, int valor)
+        public Articulo(string nombre, decimal valor)
         {
             Nombre = nombre;
             Valor = valor;
